Backend segments. Golden processor -> emulator.
