+++ tools/src/tensil/tools/compiler/OnnxFrontend.scala
@@ -385,16 +385,14 @@ class OnnxFrontend(
   ): Seq[Emitter] =
     recursiveRewrite(protos, emitter +: emitters)
 
-  private var layerIndex = 0
+  private var nextLayerIndex = 0
 
   private def startLayer(nodeProtos: Seq[NodeProto]): Scheduler = {
-    val name =
-      s"LAYER $layerIndex (${nodeProtos.map(_.name.get).mkString(",")})"
-
-    layerIndex += 1
+    val layerIndex = nextLayerIndex
+    nextLayerIndex += 1
 
     new Scheduler(
-      name,
+      layerIndex,
       arch,
       options
     )
@@ -422,7 +420,7 @@ class OnnxFrontend(
   ): Emitter =
     (context: EmitContext) => {
       if (graphPrinter.isDefined)
-        graphPrinter.get.startLayer(s"layer_$layerIndex")
+        graphPrinter.get.startLayer(s"layer_$nextLayerIndex")
 
       val scheduler = startLayer(
         Seq(
