+++ tools/test/src/tools/CompilerBackendSpec.scala
@@ -33,9 +33,10 @@ class CompilerBackendSpec extends FlatSpec {
   val Stride0Depth: Int       = 1 << 2
   val Stride1Depth: Int       = 1 << 2
 
-  def mk16BitBackend(out: OutputStream) = {
+  val SegmentKey = (0, 0, 0, 0)
+
+  def mk16BitBackend() = {
     new Backend(
-      out,
       new InstructionLayout(
         Architecture.mkWithDefaults(
           arraySize = 4,
@@ -47,14 +48,12 @@ class CompilerBackendSpec extends FlatSpec {
           stride0Depth = Stride0Depth,
           stride1Depth = Stride1Depth
         )
-      ),
-      ArchitectureDataType.FLOAT32
+      )
     )
   }
 
-  def mk32BitBackend(out: OutputStream) = {
+  def mk32BitBackend() = {
     new Backend(
-      out,
       new InstructionLayout(
         Architecture.mkWithDefaults(
           4,
@@ -66,14 +65,12 @@ class CompilerBackendSpec extends FlatSpec {
           stride0Depth = Stride0Depth,
           stride1Depth = Stride1Depth
         )
-      ),
-      ArchitectureDataType.FLOAT32
+      )
     )
   }
 
-  def mk64BitBackend(out: OutputStream) = {
+  def mk64BitBackend() = {
     new Backend(
-      out,
       new InstructionLayout(
         Architecture.mkWithDefaults(
           4,
@@ -85,20 +82,20 @@ class CompilerBackendSpec extends FlatSpec {
           stride0Depth = Stride0Depth,
           stride1Depth = Stride1Depth
         )
-      ),
-      ArchitectureDataType.FLOAT32
+      )
     )
   }
 
   it should "emit 16-bit NoOp" in {
     val out     = new ByteArrayOutputStream()
-    val backend = mk16BitBackend(out)
+    val backend = mk16BitBackend()
 
-    val stream = backend.mkStream("")
-    stream.emitNoOp()
-    stream.emitNoOp()
-    stream.emitNoOp()
-    backend.writeStream(stream)
+    val segment = backend.mkSegment(SegmentKey)
+    segment.emitNoOp()
+    segment.emitNoOp()
+    segment.emitNoOp()
+    backend.finalizeSegment(segment)
+    backend.writeSegments(out)
 
     assert(
       out.toByteArray() === Array(
@@ -120,13 +117,14 @@ class CompilerBackendSpec extends FlatSpec {
 
   it should "emit 32-bit NoOp" in {
     val out     = new ByteArrayOutputStream()
-    val backend = mk32BitBackend(out)
+    val backend = mk32BitBackend()
 
-    val stream = backend.mkStream("")
-    stream.emitNoOp()
-    stream.emitNoOp()
-    stream.emitNoOp()
-    backend.writeStream(stream)
+    val segment = backend.mkSegment(SegmentKey)
+    segment.emitNoOp()
+    segment.emitNoOp()
+    segment.emitNoOp()
+    backend.finalizeSegment(segment)
+    backend.writeSegments(out)
 
     assert(
       out.toByteArray() === Array(
@@ -148,13 +146,14 @@ class CompilerBackendSpec extends FlatSpec {
 
   it should "emit 64-bit NoOp" in {
     val out     = new ByteArrayOutputStream()
-    val backend = mk64BitBackend(out)
+    val backend = mk64BitBackend()
 
-    val stream = backend.mkStream("")
-    stream.emitNoOp()
-    stream.emitNoOp()
-    stream.emitNoOp()
-    backend.writeStream(stream)
+    val segment = backend.mkSegment(SegmentKey)
+    segment.emitNoOp()
+    segment.emitNoOp()
+    segment.emitNoOp()
+    backend.finalizeSegment(segment)
+    backend.writeSegments(out)
 
     assert(
       out.toByteArray() === Array(
@@ -176,10 +175,10 @@ class CompilerBackendSpec extends FlatSpec {
 
   it should "emit 32-bit MatMul" in {
     val out     = new ByteArrayOutputStream()
-    val backend = mk32BitBackend(out)
+    val backend = mk32BitBackend()
 
-    val stream = backend.mkStream("")
-    stream.emitMatMul(
+    val segment = backend.mkSegment(SegmentKey)
+    segment.emitMatMul(
       accumulate = false,
       0,
       MemoryAddress(MemoryTag.Local, MemoryRef.Invalid, 0x00012345L),
@@ -187,7 +186,7 @@ class CompilerBackendSpec extends FlatSpec {
       MemoryAddress(MemoryTag.Accumulators, MemoryRef.Invalid, 0x0089abcdL),
       0x543210L
     )
-    stream.emitMatMul(
+    segment.emitMatMul(
       accumulate = true,
       2,
       MemoryAddress(MemoryTag.Local, MemoryRef.Invalid, 0x00012345L),
@@ -195,7 +194,7 @@ class CompilerBackendSpec extends FlatSpec {
       MemoryAddress(MemoryTag.Accumulators, MemoryRef.Invalid, 0x0089abcdL),
       0x543210L
     )
-    stream.emitMatMul(
+    segment.emitMatMul(
       accumulate = false,
       0,
       MemoryAddress(MemoryTag.Zeroes, MemoryRef.Invalid, 0x00012345L),
@@ -203,7 +202,7 @@ class CompilerBackendSpec extends FlatSpec {
       MemoryAddress(MemoryTag.Accumulators, MemoryRef.Invalid, 0x0089abcdL),
       0x543210L
     )
-    stream.emitMatMul(
+    segment.emitMatMul(
       accumulate = true,
       0,
       MemoryAddress(MemoryTag.Zeroes, MemoryRef.Invalid, 0x00012345L),
@@ -211,7 +210,8 @@ class CompilerBackendSpec extends FlatSpec {
       MemoryAddress(MemoryTag.Accumulators, MemoryRef.Invalid, 0x0089abcdL),
       0x543210L
     )
-    backend.writeStream(stream)
+    backend.finalizeSegment(segment)
+    backend.writeSegments(out)
 
     assert(
       out.toByteArray() === Array(
@@ -253,10 +253,10 @@ class CompilerBackendSpec extends FlatSpec {
 
   it should "emit 32-bit SIMD" in {
     val out     = new ByteArrayOutputStream()
-    val backend = mk32BitBackend(out)
+    val backend = mk32BitBackend()
 
-    val stream = backend.mkStream("")
-    stream.emitSIMD(
+    val segment = backend.mkSegment(SegmentKey)
+    segment.emitSIMD(
       accumulate = false,
       SIMDOp.Zero,
       0,
@@ -265,7 +265,7 @@ class CompilerBackendSpec extends FlatSpec {
       MemoryAddress(MemoryTag.Invalid, MemoryRef.Invalid, 0x00012345L),
       MemoryAddress(MemoryTag.Invalid, MemoryRef.Invalid, 0x0089abcdL)
     )
-    stream.emitSIMD(
+    segment.emitSIMD(
       accumulate = false,
       SIMDOp.Max,
       SIMDSource.Input,
@@ -274,7 +274,8 @@ class CompilerBackendSpec extends FlatSpec {
       MemoryAddress(MemoryTag.Accumulators, MemoryRef.Invalid, 0x00012345L),
       MemoryAddress(MemoryTag.Accumulators, MemoryRef.Invalid, 0x0089abcdL),
     )
-    backend.writeStream(stream)
+    backend.finalizeSegment(segment)
+    backend.writeSegments(out)
 
     assert(
       out.toByteArray() === Array(
@@ -304,11 +305,11 @@ class CompilerBackendSpec extends FlatSpec {
 
   it should "emit 32-bit DataMove" in {
     val out     = new ByteArrayOutputStream()
-    val backend = mk32BitBackend(out)
+    val backend = mk32BitBackend()
 
-    val stream = backend.mkStream("")
+    val segment = backend.mkSegment(SegmentKey)
 
-    stream.emitDataMove(
+    segment.emitDataMove(
       toLocal = true,
       accumulate = false,
       0,
@@ -317,7 +318,7 @@ class CompilerBackendSpec extends FlatSpec {
       MemoryAddress(MemoryTag.Vars, MemoryRef.Invalid, 0x0089abcdL),
       0x00543210L
     )
-    stream.emitDataMove(
+    segment.emitDataMove(
       toLocal = false,
       accumulate = false,
       2,
@@ -326,7 +327,7 @@ class CompilerBackendSpec extends FlatSpec {
       MemoryAddress(MemoryTag.Vars, MemoryRef.Invalid, 0x0089abcdL),
       0x00543210L
     )
-    stream.emitDataMove(
+    segment.emitDataMove(
       toLocal = true,
       accumulate = false,
       3,
@@ -335,7 +336,7 @@ class CompilerBackendSpec extends FlatSpec {
       MemoryAddress(MemoryTag.Consts, MemoryRef.Invalid, 0x0089abcdL),
       0x00543210L
     )
-    stream.emitDataMove(
+    segment.emitDataMove(
       toLocal = false,
       accumulate = false,
       1,
@@ -344,7 +345,7 @@ class CompilerBackendSpec extends FlatSpec {
       MemoryAddress(MemoryTag.Consts, MemoryRef.Invalid, 0x0089abcdL),
       0x00543210L
     )
-    stream.emitDataMove(
+    segment.emitDataMove(
       toLocal = true,
       accumulate = true,
       1,
@@ -353,7 +354,7 @@ class CompilerBackendSpec extends FlatSpec {
       MemoryAddress(MemoryTag.Accumulators, MemoryRef.Invalid, 0x0089abcdL),
       0x00543210L
     )
-    stream.emitDataMove(
+    segment.emitDataMove(
       toLocal = false,
       accumulate = false,
       2,
@@ -362,7 +363,7 @@ class CompilerBackendSpec extends FlatSpec {
       MemoryAddress(MemoryTag.Accumulators, MemoryRef.Invalid, 0x0089abcdL),
       0x00543210L
     )
-    stream.emitDataMove(
+    segment.emitDataMove(
       toLocal = false,
       accumulate = true,
       3,
@@ -371,7 +372,8 @@ class CompilerBackendSpec extends FlatSpec {
       MemoryAddress(MemoryTag.Accumulators, MemoryRef.Invalid, 0x0089abcdL),
       0x00543210L
     )
-    backend.writeStream(stream)
+    backend.finalizeSegment(segment)
+    backend.writeSegments(out)
 
     assert(
       out.toByteArray() === Array(
@@ -437,35 +439,36 @@ class CompilerBackendSpec extends FlatSpec {
 
   it should "emit 32-bit LoadWeights" in {
     val out     = new ByteArrayOutputStream()
-    val backend = mk32BitBackend(out)
+    val backend = mk32BitBackend()
 
-    val stream = backend.mkStream("")
-    stream.emitLoadWeights(
+    val segment = backend.mkSegment(SegmentKey)
+    segment.emitLoadWeights(
       0,
       MemoryAddress(MemoryTag.Local, MemoryRef.Invalid, 0x00012345L),
       0x0089abcdL
     )
-    stream.emitLoadWeights(
+    segment.emitLoadWeights(
       1,
       MemoryAddress(MemoryTag.Local, MemoryRef.Invalid, 0x00012345L),
       0x0089abcdL
     )
-    stream.emitLoadWeights(
+    segment.emitLoadWeights(
       2,
       MemoryAddress(MemoryTag.Local, MemoryRef.Invalid, 0x00012345L),
       0x0089abcdL
     )
-    stream.emitLoadWeights(
+    segment.emitLoadWeights(
       3,
       MemoryAddress(MemoryTag.Local, MemoryRef.Invalid, 0x00012345L),
       0x0089abcdL
     )
-    stream.emitLoadWeights(
+    segment.emitLoadWeights(
       0,
       MemoryAddress.Zeroes,
       0x0089abcdL
     )
-    backend.writeStream(stream)
+    backend.finalizeSegment(segment)
+    backend.writeSegments(out)
 
     assert(
       out.toByteArray() === Array(
