+++ tools/test/src/tools/MaxPool.scala
@@ -6,7 +6,7 @@ package tensil.tools
 import java.io._
 import scala.reflect.ClassTag
 import tensil.tools.data.{Shape, Tensor}
-import tensil.tools.golden.{Processor, ExecutiveTraceContext}
+import tensil.tools.emulator.{Emulator, ExecutiveTraceContext}
 import tensil.ArchitectureDataType
 
 object MaxPool {
