+++ tools/test/src/tools/ResNet.scala
@@ -6,7 +6,7 @@ package tensil.tools
 import java.io._
 import scala.reflect.ClassTag
 import scala.io.Source
-import tensil.tools.golden.{Processor, ExecutiveTraceContext}
+import tensil.tools.emulator.{Emulator, ExecutiveTraceContext}
 import tensil.ArchitectureDataType
 
 object ResNet {
