+++ tools/test/src/tools/TinyYolo.scala
@@ -5,7 +5,7 @@ package tensil.tools
 
 import java.io._
 import scala.reflect.ClassTag
-import tensil.tools.golden.{Processor, ExecutiveTraceContext}
+import tensil.tools.emulator.{Emulator, ExecutiveTraceContext}
 import scala.collection.mutable
 import scala.io.Source
 import tensil.ArchitectureDataType
